backend/web/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
PyJWT==2.3.0
pycodestyle==2.8.0
toml==0.10.2